+++ code/sys/local.h
@@ -81,6 +81,40 @@ unsigned int CON_LogWrite( const char *in );
 unsigned int CON_LogRead( char *out, unsigned int outSize );
 
 
+typedef struct
+{
+	void *OpenGLLib; // instance of OpenGL library
+	FILE *log_fp;
+
+	int	monitorCount;
+
+	qboolean gammaSet;
+
+	qboolean cdsFullscreen;
+
+	qboolean dga_ext;
+
+	qboolean vidmode_ext;
+	qboolean vidmode_active;
+	qboolean vidmode_gamma;
+
+	qboolean randr_ext;
+	qboolean randr_active;
+	qboolean randr_gamma;
+
+	qboolean desktop_ok;
+	int desktop_width;
+	int desktop_height;
+	int desktop_x;
+	int desktop_y;
+} glwstate_t;
+
+
+typedef struct sym_s
+{
+	void **symbol;
+	const char *name;
+} sym_t;
 
 #ifdef MACOS_X
 char *Sys_StripAppBundle( char *pwd );
